pyproject.toml
[project]
name = "projet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docstring-parser>=0.16",
    "haystack-ai>=2.12.2",
    "huggingface-hub>=0.30.2",
    "mistral-haystack>=0.1.1",
    "pandas>=2.2.3",
    "pgvector-haystack>=3.2.0",
    "psycopg2>=2.9.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "qdrant-haystack>=9.1.1",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.40",
]

[tool.uv.sources]
huggingface-hub = { git = "https://github.com/huggingface/huggingface_hub" }
